+++ examples/data-sources/nxos_feature_macsec/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_feature_macsec" "example" {
+}