+++ examples/resources/nxos_feature_macsec/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_macsec.example "sys/fm/macsec"