+++ examples/resources/nxos_feature_macsec/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_feature_macsec" "example" {
+  admin_state = "disabled"
+}